git commit -m "Rectangle with double"

+++ guru-shifu/Rectangle/src/main/java/com/thoughtworks/rectangle/Rectangle.java
@@ -1,12 +1,12 @@
 public class Rectangle {
-    private final int length,breadth;
+    private final double length,breadth;
 
-    Rectangle(int length,int breadth) {
+    Rectangle(double length,double breadth) {
         this.length = length;
         this.breadth= breadth;
     }
 
-    public int area(){
+    public double area(){
         return length*breadth;
     }
 }
